+++ src/App.vue
@@ -7,7 +7,7 @@ themeStore.init();
 </script>
 
 <template>
-    <main class="w-full bg-zinc-200 dark:bg-zinc-950">
+    <main class="h-full w-full bg-zinc-200 dark:bg-zinc-950">
         <Toast class="max-w-xs sm:max-w-full"/>
         <RouterView />
     </main>